requirements.txt
python-telegram-bot==20.7
openai==1.12.0
sentence-transformers==2.2.2
chromadb==0.4.22
tiktoken==0.5.2
python-dotenv==1.0.0
pypdf==3.17.4
numpy==1.24.3
torch==2.1.2
fastapi==0.104.1
uvicorn==0.24.0
